+++ CHANGELOG.md
@@ -3,13 +3,18 @@
 The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project loosely adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+Format: `[MAJOR.MINOR.PATCH] - YYYY-MM-DD`
 
 - todo :: Code refactoring via `CTRL`+`ALT`+`L`
 - todo :: Add working file structure
 - todo :: Add code refactoring
 - fixme :: See error in config.cpp example, where there is an issue with < in a Q() macro and also !
 - fixme :: Consider fixing classnames that start with a number, seems like they are broken, check `30Rnd_556x45_Stanag`
+
+## [Unreleased]
+
+### Fixed
+- Support for classnames starting with numbers, like `30Rnd_556x45_Stanag`
 
 
 ## [1.0.1] - 2025-12-03